Improve Map get/remove performance
Rather than keeping a list of pairs and doing a linear search any time we need
to find a key, use a hashing strategy similar to what the extension uses in
order to maintain a lookup table by key hash. Maintain a parallel ordered list
for operations that need to respect order (e.g., iteration, reverse, sort, and
so on). In order to keep the two parallel structure in sync, wrap each pair in
a "PairRef" that keeps track of the pair's position in the internal ordered
list, and update the PairRefs when changing the order (e.g., after a sort).