Map: Only use array for collision list of size 2+
Arrays eat up a lot of memory, and there's also overhead involved in
initializing, growing, and iterating over them. Since we hope our hashing
strategy will result in relatively few collisions, we can avoid a lot of array
allocation and iteration by first associating a PairRef with a hash key, only
"upgrading" to an array once we have an actual collision. We pay in terms of
complexity because insertion, lookup, and deletion have to handle the two cases
for a value associated with a hash key: a PairRef or an array of PairRefs.

After making this change and running the benchmark, it appears that this more
complex approach is somewhat faster and about half as memory-intensive as the
previous (always use an array) approach. The main goal here is to reduce memory
usage for relatively small data sets (up to a few thousand values), so this
seems like a definite win.