Further reduce Map memory usage
Rather than wrapping a Pair in a PairRef, capture the relevant Pair
functionality in PairRef and add the ability to export to a Pair. This saves
a bit on Map storage because we now create one PairRef per value instead of
both a Pair and a PairRef, and it's a bit faster, presumably because we're
allocating less and doing less dereferencing to get at pair keys/values.

It's relatively rare that we actually want to get Pair instances out of the
collection, and in the cases where we do, we always want a copy. Previously,
some methods (e.g., first()) were returning a Pair directly instead of a copy,
but I believe that was a mistake. I think the extension makes a new Pair with
the internal key and value in these cases, and that strikes me as more
reasonable.