Changed the docblock summary for push()
The summary stated that push() would **push** the values **into the front** of the queue.
The php.net docs and wikipedia (linked from the ds-ext blog post) state that values are *pushed onto the back* of the queue and *popped off the front* of the queue.
I think it should either:
- push values into the queue from the back OR
- push values onto the back of the queue

I hope my understanding is correct and the explanation makes sense. :S

If this is right, I think we should also update the class docblock summary to mention that values are pushed onto the back.

A “first in, first out” or “FIFO” collection that pushes new values into the back of the queue but only allows access to the value at the front of the queue and iterates in order from the front, destructively.